Discard search requests in CA client
Search requests could accidentally reach a CA client if an IOC
is configured with `EPICS_CA_SERVER_PORT=5065`.

This change filters those messages to prevent parsing a search request
as a search reply and receiving several log with
"Identical process variable names on multiple servers" in which the
address is invalid.