Support full static linking in CONFIG_COMMON.
For now, only linuxCommon has the necessary definitions for this to
work. Users must define FULL_STATIC_BUILD in CONFIG_SITE for their IOC.
We don't add a new value for STATIC_BUILD since it's used in multiple
places and would make any changes to the build system more complicated.