IPv6: handle [%en0] in aToIPAddr.c
Prepare to be able to have EPICS_CAS_INTF_ADDR_LIST or
EPICS_CAS_BEACON_ADDR_LIST to be populated with "network addresses":
The official (and long) form would be something like this:
[fe80:0:0:0:cdc:863f:2e6e:b4c7%eth0]

This change prepare a shorter form:
[%eth0]

However, the logic for EPICS_CAS_AUTO_BEACON_ADDR_LIST
in caservertask.c and needs to be reviewed for the IPv6 changes.
That is for the next commits.