IPv6: Fix double epicsSocketDestroy() in repeater.cpp
When creating a socket and trying to bind to a port, it may happen
that the bind() fails. As a result, the current code called
epicsSocketDestroy() twice, whichj is simply wrong.

Detected (and fixed) another problem:
If sock6 is INVALID, don't add it to the pollFds (and produce a crash)