IPv6: Create sIPv6 sockets with ipv6only
More preparation to separation of IPv4 and IPv6 sockets.
By default, FreeBSD does not receive IPv4 on an IPv6 socket
(in opposite to Linux and Mac Os), ipv6only is 1 by default.
Even if this may be changed by the system operator, do not rely on
this in general:
Without furhther testing, I would suspect the e.g. VxWorks may have the
same philosophy.
For the moment, open 2 sockets, when IPv4 and IPv6 are wanted.
This allows more cleanups in later commits.

The only good motivation I could found is here:
https://security.stackexchange.com/questions/92081/is-receiving-ipv4-connections-on-af-inet6-sockets-insecure