iocsh: allow history and history file size to be limited
Adds EPICS_IOCSH_HISTSIZE and EPICS_IOCSH_HISTFILESIZE

EPICS_IOCSH_HISTSIZE defines the number of history items to keep in
memory, limited by stifle_history. Values of < 0 mean history is
unlimited. This is not recommended for obvious reasons. Non-numeric
values are simply ignored and the default is used instead.

EPICS_IOCSH_HISTFILESIZE defines the number of history items to store on
disk. Values < 0 or non-numeric values remove any limit on this file's
size. Values >= 0 truncate the history file to that many lines. Note
that a value of 0 will simply generate an empty history file.

This behavior mirrors bash's HISTFILESIZE and HISTSIZE.