Allow RTEMS non epics task to set task priority
If RTEMS is used with the POSIX API, the init task
'POSIX_Init()' is unfortunately given the priority '2'. This
corresponds to the second lowest POSIX prio (RTEMS pthread
prio 256). In order to give the 'rtems-init-task' the appropriate
IOCsh prio a call of 'epicsThreadSetPriority()' is necessary:
'epicsThreadSetPriority(epicsThreadGetIdSelf(), epicsThreadPriorityIocsh);'
However, 'POSIX_Init()' itself is not an epics task.