API to renumber socket FDs for code that uses select() and libCa
The fd assigned to a new socket is normally the lowest fd available.
The select() system call can only manage sockets where fd < FD_SETSIZE.
The EPICS PCAS uses select(), so can only serve clients that come in on
low-numbered sockets. The CA Gateway calls libCa as well as PCAS, so to
maximize the clients PCAS can serve, sockets created by libCa must be
allocated an fd of FD_SETSIZE or above.

On systems that support the fcntl() F_DUPFD_CLOEXEC this commit adds a
routine epicsSocketRenumber() to osiSock.h that renumbers an fd to one
that's >= FD_SETSIZE.

The CPP macro HAS_SOCK_RENUMBER will be defined after including the
osiSock.h header on systems where this API exists.