caget: convert INT to SHORT in dbr type.
Despite being documented in the caget help message as possible arguments
for the '-d' option, passing DBR*_INT types to caget using their names
(instead of numerical values) does not work.

This change makes the interface consistent with the help message, which
documents DBR*_INT types as separate items, though they are equivalent
to SHORT types.

The INT to SHORT replacement happens before any calls to
dbr_text_to_type, because it's impossible for a type ending in "INT" to
be parsed successfully by dbr_text_to_type. And the longest DBR type
name is shorter than 20 chars, so 30 chars is definitely a safe bet for
temporary buffers (and matches the size used elsewhere in the function).

We rename the 'str' buffer to 'dbrStr' to help with disambiguation.