Add `coverage` target
This target runs `lcov` to analyse coverage, and then `genhtml` to produce
a HTML report of coverage data.

The new target cannot depend on `install` and `runtests` since `runtests` is
defined in `RULES_DIRS`, so we instead perform a check to ensure that the
user has performed the appropriate actions (build and link with `--coverage`,
and run test suite) ahead of time.