Add a variable for handling the C/C++ standard
There is no standard location to specify a C/C++ standard in a way
that doesn't cause some potential conflict; ARCH_DEP_C*FLAGS seems
the most logical, but we pass the CFLAGS in that case to the CXXFLAGS
which cases warnings. POSIX is the next best, but if we are not in the
posix case it would be morally incorrect.